Simplify the identity policy methods again and not use Reg.
While the actual calling sites are in fact a bit more involved than the
Reg versions, the registration code is significantly less complicated --
the identity policy was always a weird one there anyway.
