+++ morepath/authentication.py
@@ -50,28 +50,6 @@ class IdentityPolicyRegistry(object):
         self.setting_registry = setting_registry
         self.identity_policy = None
 
-    def register_identity_policy_function(self, factory, dispatch, name):
-        """Register a method from the identity policy as a function.
-
-        The identity policy is registered as a class, but their methods
-        are really registered with dispatch functions that are then
-        exposed to the public API and are used in the framework.
-
-        :param factory: factory to create identity policy instance,
-          typically the identity policy class object.
-        :param dispatch: the dispatch function we want to register
-          a method on.
-        :param name: the name of the method to register.
-        """
-        # make sure we only have a single identity policy
-        identity_policy = self.identity_policy
-        if identity_policy is None:
-            self.identity_policy = identity_policy = mapply(
-                factory,
-                settings=self.setting_registry)
-        func = getattr(identity_policy, name)
-        self.reg_registry.register_function(dispatch, func)
-
 
 class Identity(object):
     """Claimed identity of a user.
